# Add LoongArch64 sha1 assembly support