@newpavlov 2023-09-19
Hm, the performance difference is relatively small, but well above noise levels, so I guess we can merge this PR. You could squeeze a bit more performance by implementing the looping over blocks inside assembly. It will allow you to keep hasher state in registers without spilling it into memory after processing each block.

Note that the asm-hashes repository crates are  effectively deprecated and we do not plan to use them in the next release cycle of the hashes repository crates. If you have time, it would be better to adapt this PR to using `asm!` directly inside the `sha1` crate.