@heiher 2023-09-21
Completed in PR: https://github.com/RustCrypto/hashes/pull/504